backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pymsgbox==1.0.9
PyPDF2==3.0.1
extract-msg==0.41.5
Pillow==10.0.1
python-magic==0.4.27
Wand==0.6.11
olefile==0.46
pdfkit==1.0.0
reportlab==4.0.4
pytest==7.4.2
Flask-Cors==4.0.0
pytest-cov==4.1.1
pytest-mock==3.11.1
gunicorn==21.2.0 